Unique options derive cell values.
When a unique option is found among cells, the cell carrying the unique option is derived to that value
