+++ src/app/solver/sudoku-solver-input-cell/sudoku-solver-input-cell.component.scss
@@ -17,6 +17,7 @@ ul.options-list {
   flex-direction: row;
   flex-wrap:wrap;
   list-style-type: none;
+  font-size: 0.8rem;
 
 }
